+++ src/Accounts/Accounts/Environment/SetAzureRMEnvironment.cs
@@ -12,14 +12,10 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
-using Microsoft.Azure.Commands.Common.Authentication.Models;
-using Microsoft.Azure.Commands.Common.Authentication.ResourceManager;
 using Microsoft.Azure.Commands.Profile.Common;
 using Microsoft.Azure.Commands.Profile.Models;
 using Microsoft.Azure.Commands.Profile.Utilities;
-using Microsoft.Azure.Commands.ResourceManager.Common;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System;
 using System.Globalization;
@@ -198,7 +194,7 @@ namespace Microsoft.Azure.Commands.Profile
                     if (AzureEnvironment.PublicEnvironments.Keys.Any((k) => string.Equals(k, Name, StringComparison.CurrentCultureIgnoreCase)))
                     {
                         throw new InvalidOperationException(string.Format(CultureInfo.CurrentCulture,
-                            "Cannot change built-in environment {0}.", Name));
+                            "Cannot change built-in or discovered environment {0}.", Name));
                     }
 
                     if (this.ParameterSetName.Equals(MetadataParameterSet, StringComparison.Ordinal))
@@ -214,7 +210,11 @@ namespace Microsoft.Azure.Commands.Profile
                         IAzureEnvironment newEnvironment;
                         if (!defProfile.TryGetEnvironment(this.Name, out newEnvironment))
                         {
-                            newEnvironment = new AzureEnvironment { Name = this.Name };
+                            newEnvironment = new AzureEnvironment
+                            {
+                                Name = this.Name,
+                                Type = AzureEnvironment.TypeUserDefined
+                            };
                         }
 
                         if (publicEnvironment.Key == null)