+++ src/Websites/Websites/Cmdlets/WebApps/SetAzureWebApp.cs
@@ -96,15 +96,14 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
         public string ContainerImageName { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Private Container Registry Server Url", ParameterSetName = ParameterSet1Name)]
-        [ValidateNotNullOrEmpty]
+        [ValidateNotNull]
         public string ContainerRegistryUrl { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Private Container Registry Username", ParameterSetName = ParameterSet1Name)]
-        [ValidateNotNullOrEmpty]
+        [ValidateNotNull]
         public string ContainerRegistryUser { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Private Container Registry Password", ParameterSetName = ParameterSet1Name)]
-        [ValidateNotNullOrEmpty]
         public SecureString ContainerRegistryPassword { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Enables/Disables container continuous deployment webhook", ParameterSetName = ParameterSet1Name)]
@@ -220,18 +219,31 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
                         }
                     }
                     
-
                     if (ContainerRegistryUrl != null)
                     {
-                        appSettings[CmdletHelpers.DocerRegistryServerUrl] = ContainerRegistryUrl;
+                        appSettings.Remove(CmdletHelpers.DockerRegistryServerUrl);
+
+                        if (ContainerRegistryUrl != string.Empty)
+                        {
+                            appSettings[CmdletHelpers.DockerRegistryServerUrl] = ContainerRegistryUrl;
+                        }
                     }
+
                     if (ContainerRegistryUser != null)
                     {
-                        appSettings[CmdletHelpers.DocerRegistryServerUserName] = ContainerRegistryUser;
+                        appSettings.Remove(CmdletHelpers.DockerRegistryServerUserName);
+
+                        if (ContainerRegistryUser != string.Empty)
+                        {
+                            appSettings[CmdletHelpers.DockerRegistryServerUserName] = ContainerRegistryUser;
+                        }                        
                     }
+
+                    appSettings.Remove(CmdletHelpers.DockerRegistryServerPassword);
+
                     if (ContainerRegistryPassword != null)
                     {
-                        appSettings[CmdletHelpers.DocerRegistryServerPassword] = ContainerRegistryPassword.ConvertToString();
+                        appSettings[CmdletHelpers.DockerRegistryServerPassword] = ContainerRegistryPassword.ConvertToString();
                     }
 
                     if (parameters.Contains("EnableContainerContinuousDeployment"))