+++ src/PowerBIEmbedded/PowerBIEmbedded/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.1.2
+* Updated assembly version of PowerBI cmdlets
+
 ## Version 1.1.1
 * Update references in .psd1 to use relative path