+++ src/Security/Security/ChangeLog.md
@@ -18,6 +18,18 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add new cmdlets: 'Get-AzSecurityAssessment',
+                   'Set-AzSecurityAssessment',
+                   'Remove-AzSecurityAssessment',
+                   'Get-AzSecurityAssessmentMetadata',
+                   'Set-AzSecurityAssessmentMetadata',
+                   'Remove-AzSecurityAssessmentMetadata',
+                   'Get-AzSecuritySubAssessment'
+
+## Version 0.7.9
+* Added new cmdlets: 
+	- `Get-AzSecuritySetting`
+	- `Set-AzSecuritySetting`
 
 ## Version 0.7.8
 * Add new cmdlets: `Get-AzRegulatoryComplianceStandard`,